pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["KIMBERLY DAMALAS <kimatcamp@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
psycopg2-binary = "^2.8.6"
postgres = "^3.0.0"
docker = "^4.4.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
